hackathon-18march-boilerplate/consumer/requirements.txt
orca-platform-sdk-ui
uvicorn
fastapi
